+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Updated `setup-r` script for R 4.6
 
-* Updated and exported function `print_renv_vulns()`, which prints vulnerable packages using YAML
+* Updated and exported function `renv_vulns()`, which prints vulnerable packages using YAML
 
 # rdev 1.16.5